+++ pingche_project/src/main/java/com/yf/pingche/entity/User.java
@@ -17,7 +17,7 @@ import lombok.experimental.Accessors;
  * </p>
  *
  * @author yangfei
- * @since 2020-07-03
+ * @since 2020-07-04
  */
 @Data
 @EqualsAndHashCode(callSuper = false)
@@ -29,7 +29,7 @@ public class User implements Serializable {
 private static final long serialVersionUID=1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @TableField("avatarUrl")
     private String avatarUrl;
